+++ src/Draftor/Converters/TotalToIntConverter.cs
@@ -10,7 +10,7 @@ public class TotalToIntConverter : IValueConverter
 
     public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
     {
-        if (value is double a)
+        if (value is decimal a)
         {
             if (a > 0)
                 return 1;
